rename TableEnvironments, TableEnvironment -> AbstractTableEnvironment, GeneralTableEnvironment -> TableEnvironment
