+++ flink-libraries/flink-sql-client/src/test/java/org/apache/flink/table/client/gateway/local/ExecutionContextTest.java
@@ -23,8 +23,8 @@ import org.apache.flink.api.common.restartstrategy.RestartStrategies;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.client.cli.DefaultCLI;
 import org.apache.flink.configuration.Configuration;
+import org.apache.flink.table.api.AbstractTableEnvironment;
 import org.apache.flink.table.api.StreamTableEnvironment;
-import org.apache.flink.table.api.TableEnvironment;
 import org.apache.flink.table.api.Types;
 import org.apache.flink.table.client.config.Environment;
 import org.apache.flink.table.client.gateway.SessionContext;
@@ -73,7 +73,7 @@ public class ExecutionContextTest {
 	@Test
 	public void testFunctions() throws Exception {
 		final ExecutionContext<?> context = createDefaultExecutionContext();
-		final TableEnvironment tableEnv = context.createEnvironmentInstance().getTableEnvironment();
+		final AbstractTableEnvironment tableEnv = context.createEnvironmentInstance().getTableEnvironment();
 		final String[] expected = new String[]{"scalarUDF", "tableUDF", "aggregateUDF"};
 		final String[] actual = tableEnv.listUserDefinedFunctions();
 		Arrays.sort(expected);
@@ -119,7 +119,7 @@ public class ExecutionContextTest {
 			new TypeInformation[]{Types.BOOLEAN(), Types.STRING()},
 			sinks.get("TableSourceSink").getFieldTypes());
 
-		final TableEnvironment tableEnv = context.createEnvironmentInstance().getTableEnvironment();
+		final AbstractTableEnvironment tableEnv = context.createEnvironmentInstance().getTableEnvironment();
 
 		assertArrayEquals(
 			new String[]{"TableNumber1", "TableNumber2", "TableSourceSink", "TestView1", "TestView2"},
