+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/api/StreamTableEnvironment.scala
@@ -75,7 +75,7 @@ import _root_.scala.collection.JavaConverters._
 class StreamTableEnvironment(
     private[flink] val execEnv: StreamExecutionEnvironment,
     config: TableConfig)
-  extends TableEnvironment(config) {
+  extends AbstractTableEnvironment(config) {
 
   def this(env: ScalaStreamExecEnv, config: TableConfig) {
     this(env.getWrappedStreamExecutionEnvironment, config)
@@ -109,8 +109,8 @@ class StreamTableEnvironment(
     "_DataStreamTable_" + nameCntr.getAndIncrement()
 
   /**
-    * Registers an internal [[StreamTableSource]] in this [[TableEnvironment]]'s catalog without
-    * name checking. Registered tables can be referenced in SQL queries.
+    * Registers an internal [[StreamTableSource]] in this [[AbstractTableEnvironment]]'s
+    * catalog without name checking. Registered tables can be referenced in SQL queries.
     *
     * @param name        The name under which the [[TableSource]] is registered.
     * @param tableSource The [[TableSource]] to register.
@@ -207,7 +207,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers an external [[TableSink]] with given field names and types in this
-    * [[TableEnvironment]]'s catalog.
+    * [[AbstractTableEnvironment]]'s catalog.
     * Registered sink tables can be referenced in SQL DML statements.
     *
     * Example:
@@ -250,7 +250,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers an external [[TableSink]] with already configured field names and field types in
-    * this [[TableEnvironment]]'s catalog.
+    * this [[AbstractTableEnvironment]]'s catalog.
     * Registered sink tables can be referenced in SQL DML statements.
     *
     * @param name The name under which the [[TableSink]] is registered.
@@ -512,7 +512,7 @@ class StreamTableEnvironment(
   }
 
   /**
-    * Registers a [[DataStream]] as a table under a given name in the [[TableEnvironment]]'s
+    * Registers a [[DataStream]] as a table under a given name in the [[AbstractTableEnvironment]]'s
     * catalog.
     *
     * @param name The name under which the table is registered in the catalog.
@@ -534,7 +534,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers a [[DataStream]] as a table under a given name with field names as specified by
-    * field expressions in the [[TableEnvironment]]'s catalog.
+    * field expressions in the [[AbstractTableEnvironment]]'s catalog.
     *
     * @param name The name under which the table is registered in the catalog.
     * @param dataStream The [[DataStream]] to register as table in the catalog.
@@ -1070,7 +1070,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers the given [[DataStream]] as table in the
-    * [[TableEnvironment]]'s catalog.
+    * [[AbstractTableEnvironment]]'s catalog.
     * Registered tables can be referenced in SQL queries.
     *
     * The field names of the [[Table]] are automatically derived
@@ -1088,7 +1088,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers the given [[DataStream]] as table with specified field names in the
-    * [[TableEnvironment]]'s catalog.
+    * [[AbstractTableEnvironment]]'s catalog.
     * Registered tables can be referenced in SQL queries.
     *
     * Example:
@@ -1401,7 +1401,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers the given [[DataStream]] as table in the
-    * [[TableEnvironment]]'s catalog.
+    * [[AbstractTableEnvironment]]'s catalog.
     * Registered tables can be referenced in SQL queries.
     *
     * The field names of the [[Table]] are automatically derived
@@ -1419,7 +1419,7 @@ class StreamTableEnvironment(
 
   /**
     * Registers the given [[DataStream]] as table with specified field names in the
-    * [[TableEnvironment]]'s catalog.
+    * [[AbstractTableEnvironment]]'s catalog.
     * Registered tables can be referenced in SQL queries.
     *
     * Example:
