+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/api/scala/TableFunctionConversions.scala
@@ -19,7 +19,7 @@
 package org.apache.flink.table.api.scala
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.table.api.{Table, TableEnvironment}
+import org.apache.flink.table.api.{Table, AbstractTableEnvironment}
 import org.apache.flink.table.expressions._
 import org.apache.flink.table.functions.TableFunction
 import org.apache.flink.table.plan.logical.LogicalTableFunctionCall
@@ -42,7 +42,8 @@ class TableFunctionConversions[T](tf: TableFunction[T]) {
     val resultType = if (tf.getResultType == null) typeInfo else tf.getResultType
 
     new Table(
-      tableEnv = null.asInstanceOf[TableEnvironment], // Table environment will be set later.
+      // Table environment will be set later.
+      tableEnv = null.asInstanceOf[AbstractTableEnvironment],
       LogicalTableFunctionCall(
         tf.getClass.getCanonicalName,
         tf,
