+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/catalog/ExternalCatalogSchema.scala
@@ -24,7 +24,7 @@ import java.util.{Collection => JCollection, Collections => JCollections, Linked
 import org.apache.calcite.linq4j.tree.Expression
 import org.apache.calcite.rel.`type`.RelProtoDataType
 import org.apache.calcite.schema._
-import org.apache.flink.table.api.{CatalogNotExistException, TableEnvironment, TableNotExistException}
+import org.apache.flink.table.api.{CatalogNotExistException, AbstractTableEnvironment, TableNotExistException}
 import org.apache.flink.table.util.Logging
 
 import scala.collection.JavaConverters._
@@ -40,7 +40,7 @@ import scala.collection.JavaConverters._
   * @param catalog           external catalog
   */
 class ExternalCatalogSchema(
-    tableEnv: TableEnvironment,
+    tableEnv: AbstractTableEnvironment,
     catalogIdentifier: String,
     catalog: ExternalCatalog) extends Schema with Logging {
 
@@ -125,7 +125,7 @@ object ExternalCatalogSchema {
     * @param externalCatalog           The external catalog to register
     */
   def registerCatalog(
-      tableEnv: TableEnvironment,
+      tableEnv: AbstractTableEnvironment,
       parentSchema: SchemaPlus,
       externalCatalogIdentifier: String,
       externalCatalog: ExternalCatalog): Unit = {
