+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/utils/UserDefinedFunctionUtils.scala
@@ -34,7 +34,7 @@ import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.common.typeutils.CompositeType
 import org.apache.flink.api.java.typeutils.{PojoField, PojoTypeInfo, TypeExtractor}
 import org.apache.flink.table.api.dataview._
-import org.apache.flink.table.api.{TableEnvironment, TableException, ValidationException}
+import org.apache.flink.table.api.{AbstractTableEnvironment, TableEnvironment, TableException, ValidationException}
 import org.apache.flink.table.calcite.FlinkTypeFactory
 import org.apache.flink.table.dataview._
 import org.apache.flink.table.expressions._
@@ -775,7 +775,7 @@ object UserDefinedFunctionUtils {
     * @return A LogicalTableFunctionCall.
     */
   def createLogicalFunctionCall(
-      tableEnv: TableEnvironment,
+      tableEnv: AbstractTableEnvironment,
       udtf: String): LogicalTableFunctionCall = {
 
     var alias: Option[Seq[String]] = None
