+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/plan/logical/LogicalNode.scala
@@ -20,7 +20,7 @@ package org.apache.flink.table.plan.logical
 import org.apache.calcite.rel.RelNode
 import org.apache.calcite.tools.RelBuilder
 import org.apache.flink.table.plan.TreeNode
-import org.apache.flink.table.api.{TableEnvironment, ValidationException}
+import org.apache.flink.table.api.{AbstractTableEnvironment, ValidationException}
 import org.apache.flink.table.expressions._
 import org.apache.flink.table.typeutils.TypeCoercion
 import org.apache.flink.table.validate._
@@ -49,7 +49,7 @@ import org.apache.flink.table.validate._
 abstract class LogicalNode extends TreeNode[LogicalNode] {
   def output: Seq[Attribute]
 
-  def resolveExpressions(tableEnv: TableEnvironment): LogicalNode = {
+  def resolveExpressions(tableEnv: AbstractTableEnvironment): LogicalNode = {
     // resolve references and function calls
     val exprResolved = expressionPostOrderTransform {
       case u @ UnresolvedFieldReference(name) =>
@@ -79,7 +79,7 @@ abstract class LogicalNode extends TreeNode[LogicalNode] {
 
   protected[logical] def construct(relBuilder: RelBuilder): RelBuilder
 
-  def validate(tableEnv: TableEnvironment): LogicalNode = {
+  def validate(tableEnv: AbstractTableEnvironment): LogicalNode = {
     val resolvedNode = resolveExpressions(tableEnv)
     resolvedNode.expressionPostOrderTransform {
       case a: Attribute if !a.valid =>
@@ -103,7 +103,9 @@ abstract class LogicalNode extends TreeNode[LogicalNode] {
     * Resolves the given strings to a [[NamedExpression]] using the input from all child
     * nodes of this LogicalPlan.
     */
-  def resolveReference(tableEnv: TableEnvironment, name: String): Option[NamedExpression] = {
+  def resolveReference(
+      tableEnv: AbstractTableEnvironment,
+      name: String): Option[NamedExpression] = {
     // try to resolve a field
     val childrenOutput = children.flatMap(_.output)
     val fieldCandidates = childrenOutput.filter(_.name.equalsIgnoreCase(name))
